functions/requirements.txt
firebase_functions~=0.1.0
instructor
google-analytics-data